feat: Remove all cargo lockfiles when bumping the version
Cargo lockfiles also store information on versions of the subcrates.
When bumped, these versions cannot be found anymore and cargo fails
to build or test the current code.

A better solution would be to manually run cargo update -p <...> for
all subcrates automatically in all folders where a Cargo.lock is found