fix(debuginfo): Do not assume sorted compilation units (#214)
The `Object` interface was built on the invalid assumption that
compilation units are generally sorted by non-overlapping address
ranges. In fact, they are often overlapping and functions can be
declared in any order.

This patch changes the guarantees and no longer sorts `functions()`
returned by the `DebugSession` iterator. This will allow to further
optimize this iterator in the future and remove some intermediate
vectors.

The `SymcacheWriter` is now more resilient against this when merging
the symbol table into written functions.