fix(debuginfo): Skip eliminated functions in linked objects (#216)
In regular shared objects or executables, functions eliminated by the
linker have no ranges and `low_pc`. As the primary use case of this
library is to symbolicate stack traces, yielding such functions from the
function iterator is not desirable. Also, the virtual address of the
functions is no longer valid and implicitly overlaps with other
eliminated functions. By default, they should not be included.

For relocatable objects, this does not apply since DW_AT_low_pc is
always zero. In this case, do not omit such functions.