fix(symcache): Use same instruction address for full lookup (#226)
Ensures that for a lookup with inline frames the same instruction
address is yielded for every frame instead of the caller function
address. This allows Sentry to display proper inline markers.

When this was implemented originally, this changed address was supposed
to show the instruction of the call in the parent. However, this can
also be observed using `symbol_addr` on the inlinee. Having the same
instruction address is technically more correct.