fix(cfi): Detect stub DLLs and skip CFI generation (#233)
Packers may strip original sections from PEs and replace them with
stubs. In such a case, the exception table still contains references to
sections like `.pdata`, even though they are in fact empty.

To detect this case, we scan for `.stub` sections. If such a section is
present and `.pdata` is empty, then we report unwind info as missing.
This check is extremely conservative so that it does not cause false
positives.