fix(debuginfo): Detect mangled anonymous namespaces in PDB inlinees (#261)
ID records specify the mangled format for anonymous namespaces:
`?A0x<id>`, where `id` is a hex identifier of the namespace. Demanglers
usually resolve this as "anonymous namespace".

This also fixes an issue with demangling: If the anonymous namespace is
in the beginning of the composed name, it leads the demangler to believe
that this is a mangled MSVC name. However, since this is a composed
demangled name, this leads to a demangle error.