fix(symcache): Support lookup for public syms larger than 65k (#320)
Public symbol records often have no size. If there's a gap larger than
65k, the symcache lookup would not resolve the last symbol. For example,
with this Breakpad symbol a lookup for 0x180000 should resolve
`symbol_a`:

    PUBLIC 100000 0 symbol_a
    PUBLIC 200000 0 symbol_b

In practice, it is likely that there are symbols missing and the lookup
would in fact be wrong. However, there could be the possibility of such
large functions and with how symbol tables are structured, this would be
the expected result. Note that this now also matches Breakpad's
behavior.