fix(debuginfo): Detect more hidden symbols in swift sources (#316)
A regression in #127 broke detection of BCSymbolMaps. The RegEx
`__?hidden#\d+_` was replaced with a literal `__?hidden#` prefix and
missed that `?` is an operator rather than a literal question mark.

This RegEx was required due to another issue in the MachO symbol
iterator, which unconditionally strips the first leading underscore to
match names from DWARF debugging information. For hidden symbols, this
does not make sense, however, and instead it should be retained.