ref(compiler): Upgrade to Rust 1.51 (#351)
Fix some clippy warnings and disable some that we can't or want
to fix, because they're from a third party or would break the API.

The Into -> From changes are possible since the orphan rules have
been relaxed in Rust 1.41. So we no longer work on older rust
versions.

Also bring the periodic CI job to the same level as the symbolicator
one. Allowing to run it on demand and continue tests when one
fails.