Delete unused and misleading fixtures. (#427)
There are fixtures at the following paths:
 - symbolic-testutils/fixtures/
 - symbolic-debuginfo/tests/fixtures/

For example, both have a crash.pdb which looks very similar on the
surface but is actually slightly different. I lost a bunch of time
because I was investigating the wrong file.

The symbolic-debuginfo tests use fixtures under symbolic-testutils.
Most of the fixtures in symbolic-debuginfo are unused. This commit
removes them.