fix(unreal): Support the UE5 crash reporter (#449)
UE5 introduced a "CR1" marker at the beginning of crash archives. It indicates
that the header at the beginning of the file is valid. Previously, that header
was merely a placeholder and a second header at the end of the file contained
valid information. Unreal4Crash::parse now supports both versions by checking
for "CR1".