ref(usym): Add test with malformed data (#539)
There's really no reason why we can't handle usyms 
where with non-UTF8 strings. We're going to defer
this work for now though, and leave a test as a breadcrumb
of sorts to hopefully remind ourselves to revisit that later.
It also indirectly asserts that the parser doesn't have
any non-UTF8 string handling baked into it (yet).