feat(usym): Write usym records that definitely have mappings to symcache (#540)
This adds basic functionality to support and convert usyms
as generated by Unity's IL2CPP into symcaches. This is
experimental and meant to be a proof of concept, so it's
hidden behind a feature flag.

This currently only focuses on records found in usyms that
directly map native code to managed code, which means
that it currently is incapable of symbolicating native code,
although it's capable of doing so with a little bit of extra
elbow grease.