ref(symcache): Add a test for mostly overlapping functions (#545)
Because we don't track file name functions, there are situations 
where the only distinguishing difference between two otherwise 
identical functions is their entry_pc. It's important to ensure that 
this doesn't change, so a test demonstrating one of those 
situations has been added to ensure that this behaviour isn't
broken.