ref(symcache): Make `comp_dir` fields obsolete
The `comp_dir` field is removed from functions and folded into
the `directory` field for files. In both cases, the raw structures
retain the actual field for binary compatibility, but it is never
written or read. The new `SymCacheConverter::normalize_path`
function should probably be optimized—currently it allocates a
number of potentially unnecessary strings.