ref(symcache): Use utf8 paths
This replaces some strings in symcace::writer and
symcace::transform with `Uft8Path{,Buf}`s from the
`camino` crate. This lets us use a path-based API,
which is useful for path normalization, while not
having to deal with non-UTF8 paths (which we don't
care about anyway since all we want to do with paths
is display them to the user).