Process INLINE_ORIGIN and INLINE directives.
Fixes #432.

The breakpad format represents lines and inlinees in the same way as
DWARF: Line records decribe locations in the "innermost" function on the
inline stack at that address, and inlinees have "call" locations.
So a similar conversion to a tree structure is performed as in the DWARF
parser.

The new test based on crash.inlines.sym isn't perfect yet; the .sym file
contains some inconsistent pieces of information for inline functions
with multiple address ranges.
Once the underlying bug has been fixed (#603), I intend to create a
correct .sym file and the test will be better.