ref(symcache): Slightly generalize function processing (#637)
Currently, `process_symbolic_function` relies on the property that all line
records of an inlinee must also be present in the parent. This relaxes this
condition so that the  inlinee's line records don't have to occur exactly in
the parent, but instead only need to be covered by line records in the parent.