feat: Use Hermes Scope lookup directly when it is available
Hermes Source Maps have scope information embedded in them already,
which we should be able to use that when available.

The sourcemap crate currently lacks looking up the scope for a
token, and only accepts a "bytecode_offset`, so it will not
work when the sourcemap has the scope extension but we have
"line/column" info instead of "bytecode".

On second thought this was an oversight in the original
design of that feature and the sourcemap crate should not
care about that detail.