fix: Speed up the SmCache SourceContext
The SourceContext is responsible to offer fast byte offset <-> line/column
conversion. Previously it was based on saving the offset of each line
and doing a byte/utf16-offset counting loop. Turns out this is not a
good idea if you never have line breaks like in the case of minified js.

Instead, this now creates an index every X chars and rather goes from there.

This cuts total "create/write SmCache" more than in half, and the
previous "resolve original names" and "convert scope index" spans that
do a lot of conversions down by 1-2 orders of magnitude.