ref: Work around gimli re-parsing Abbreviations
Abbreviations in the debug_abbrev section can be, and often are, shared between compilation units.
However, the upstream `gimli::Unit::new` eagerly parses and allocates `Abbreviations` for each `Unit`.

We have a pathological case where an extremely large 5.5G DWARF file with ~280k compilation
units and ~1500 Abbreviations is taking a very long time to process and is holding on to an
unreasonable amount of memory while doing so.
Profiling suggests a majority of time and memory is spent parsing and dropping these Abbreviations.

This workaround copies the `gimli::Unit::new` code to avoid parsing `Abbreviations` at offset 0. In the pathological case mentioned above, *all* the compilation units share the same list of `Abbreviations` at offset 0.

Runtime of symcache conversion for the mentioned pathological case goes from 13 minutes down to slightly over 1 minute, and peak memory usage from 18G down to 8G. Not great, not terrible, but a huge improvement nonetheless.