ref: Cross-compile aarch64 python pkg
This ports https://github.com/getsentry/relay/pull/1438 to symbolic.
It is using a x86_64 docker container configured for cross compiling that builds a
aarch64 symbolic binary which is then re-used when building the aarch64 wheel.
Thus it avoids running the whole cargo build in aarch64 qemu which can be extremely slow.