Fix rust-analyzer extension ID.
The old rust-analyzer extension has been removed from the
VS Code market place, so VS Code now displays a warning
when the symbolic repo is opened in the editor. This
fixes that warning.

The file indentation has changed because it seems that
this is the first time anybody has touched this file
since editor.formatOnSave was turned on.