fix(debuginfo): Split lines in more cases.
Fixes #713.

In the FunctionBuilder, which is used for the DWARF reader and for the
breakpad reader, we were already splitting lines which were extending
beyond the end of an inlinee. Now we also split lines which extend into
the beginning of the next inlinee.

This fixes all "overlapping line" warnings in the debuginfo_debug
example when run on the macos/crash.dSYM. Those overlapping lines were
caused by "self lines" which overlapped with "inlinee lines".

I've also updated crash.inlines.sym with a file that I freshly generated
using a dump_syms build which includes this symbolic fix. Now the
crash.inlines.sym file no longer has overlapping inlinees.