feat(debuginfo): Add inlinee overlap checks.
This commit adds a few checks to the debuginfo_debug example: We now
check that the lines from sibling inlinees don't overlap with each other.

This is another failure case which is currently happening in practice,
namely in the breakpad reader, and it's confusing the symcache writer.
I will fix the breakpad reader in a different PR.