fix: Completely skip over hidden sequence points
It is a bit unclear how to really treat these, however just
setting the line to `0` means we correctly symbolicate,
but then later on we apply the wrong source context lines.

By ignoring these, we do get the correct source line (and thus source context)
for a specific customer issue.