Remove `cargo update` from version-bump script (#796)
Looks like `minidump-processor` is now using symbolic as well. And we cannot bump the version
of its symbolic dependency because its an out-of-workspace package. `cargo update` in that case would
require the version to already exist in the registry.

Just remove the `cargo update`. We do not check in our lockfile so whatever updates it would do would never be committed anyway.