fix(cfi): Skip u64::MAX FDEs when converting DWARF (#868)
We have seen FDEs with an initial address of u64::MAX in user-provided DWARF files. Such FDEs will invariably fail to process because of either an address overflow error in gimli or an underflow in the length calculation in line 756. Therefore, we skip them immediately so we don't abort the processing of the entire file.