Revert "Support for DWARFv5 embedded source code extension (#849)" (#870)
This reverts the "feature part" of commit 7dc28dd (#849). The dep updates can stay.

Unfortunately opening a debug session to check for the presence of sources is prohibitively expensive in terms of memory (going from ~450MB to ~3.7GB in a Python process that calls Object.features on an example 1.3GB file). In order to support this we'll have to find a more parsimonious way to check if a file contains sources.