fix(symcache): Use saturating addition in writer
Addition wraps in release mode. For large line size values,
this resulted in ranges whose end was before their start,
which caused us to skip a function's inlinees in some cases.

Instead of using bare addition on `u64` and casting to `u32`
at the end, we now use saturating addition on `u32` directly.