fix(demangle): Unbreak GitHub Action Windows builds (#891)
The 12.13.0 release and pull request #885 specifically broke GitHub
Action based Windows builds. The reason seems to be that the Swift
library upgrade now relies on C++17 features and the GitHub Actions
Windows image uses MSVC 2022, which requires an explicit opt-in to
those. The provided -std=c++17 works on GCC and similar, but does not
have any effect on MSVC.
To fix the build, add the flag that MSVC understands for opting in to
the new standard.