ref(sourcebundle): Check UTF-8 validity memory efficiently (#890)
The current check to ensure a sourcebundle is valid UTF-8 reads the entire sourcebundle file into memory. This is inefficient for large files.

This PR introduces a UTF8Reader which wraps any reader. The UTF8Reader ensures that the stream is valid UTF8 as it is being read, while only requiring a small amount of memory (currently 8 KiB) to be allocated as a buffer.

Co-authored-by: Sebastian Zivota <loewenheim@users.noreply.github.com>