fix(symcache): Explicitly map "holes" between functions
Currently, if we look up an address that falls "between"
two functions/symbols, we will always resolve it to the
first of the two functions/symbols because each range
implicitly extends to the start of the next one. This can
lead to bad user experience when we map (invalid) addresses
to entirely nonsensical functions.

Therefore, we now explicitly insert a mapping at the end
of the function that maps to "nowhere" to denote the
function's end.