fix(symcache): Overwrite "empty" mappings for symbols (#915)
As of #897 we explicitly insert the ends of functions and symbols into symcaches to reduce bogus mappings. While this worked correctly for functions, we introduced a bug in the case of symbols that resulted in symbols not being properly inserted into symcaches.

This fixes the bug by explicitly overwriting previously inserted "empty" mappings.