ref(symcache): Make parsing backwards-compatible
This puts in some indirection to make it so that old symcache versions
remain parseable, even if the on-disk format changes. For the two
currently supported versions, 7 and 8, this is massive overkill because
they only differ in how string lengths are serialized to the string
table. The intention is to enable us to add data to the format without
having to throw out every old symcache immediately.